feat: Add solution for finding the maximum height of a triangle
This commit adds a solution for finding the maximum height of a triangle given the number of red and blue blocks. The code calculates the maximum height by iterating through possible heights and checking if the number of red and blue blocks can form a triangle with that height. The solution is optimized by using a binary search algorithm.